requirements.txt
tqdm
pyyaml
numpy 
scipy 
pandas
matplotlib
ftfy
regex